dom/u2f/tests/test_u2f_replaceable.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Make sure that pages can define a variable named "u2f" and have it work</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    var u2f = 5;
    is(u2f, 5, "Should have allowed reassignment");
    is(Object.getOwnPropertyDescriptor(window, "u2f").value, 5,
       "Should be a value property");
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
